requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
cycler==0.10.0
dj-database-url==0.4.2
Django==1.11.7
django-extensions==2.1.2
Fabric3==1.14.post1
gunicorn==19.9.0
idna==2.6
matplotlib==2.1.0
numpy==1.13.3
pandas==0.20.2
paramiko==2.4.1
psycopg2==2.7.3.2
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
six==1.11.0
sklearn==0.0
urllib3==1.22
whitenoise==3.3.1
